requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnc==1.1.0
filelock==3.6.0
flann==1.6.13
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.4.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.11.2
joblib==1.1.0
Markdown==3.3.6
multidict==6.0.2
nltk==3.7
numpy==1.22.2
oauthlib==3.2.0
omegaconf==2.1.1
packaging==21.3
portalocker==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.7
pytorch-lightning==1.5.10
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rouge-score==0.0.4
rsa==4.8
sacrebleu==2.0.0
sacremoses==0.0.47
six==1.16.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.11.6
torch==1.10.2
torchmetrics==0.7.2
tqdm==4.63.0
transformers==4.16.2
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0
